Execute tensor product functions with RowFn
