+++ vortex-tensor/src/scalar_fns/tests/mod.rs
@@ -3,5 +3,7 @@
 
 //! Tests for the tensor scalar functions.
 
+mod cosine_similarity;
+mod inner_product;
 mod l2_norm;
 mod row;
